requirements.txt
pika
pydantic
python-dotenv
fastapi[all]
